artmintflow/requirements.txt

fastapi
uvicorn
pydantic
python-multipart
aiohttp
requests
python-dotenv
web3
